==> scripts/02_volcanoplots.Rmd <==
---
title: "volcanos"
output: md_document
---

These are the packages I need for my volcano plots.

```{r setup, include=FALSE}
library(tidyverse)
library(cowplot)
library(ggrepel)

#install.packages("ggrepel")
# set output file for figures 
knitr::opts_chunk$set(fig.path = '../figures/02_volcanoplots/')
```

Here I load the relevant dataframes and set the color palette.

```{r loaddata}
# dissocation DEGes
dissociation <- read.csv("../results/volcanoTreatment.csv", header = T, row.names = 1)

# subfield specific degs
subfield <- read.csv("../results/volcanoCA1DG.csv", header = T, row.names = 1)
head(subfield)
```

Treatement volcanos

```{r plot}
levels(dissociation$direction)

# set factor levels
dissociation$direction <- factor(dissociation$direction,
                             levels = c("HOMO", "DISS", "neither"))
#set colors
dissociationcolor <-  c("HOMO" = "black", "DISS" = "black", "neither" = "#d9d9d9")

# for comparison to heat map of 30 degs
volcanoplot <- ggplot(dissociation, aes(x = lfc, y = pvalue)) + 
  geom_point(aes(color = factor(direction), shape = factor(direction)), 
             size = 1.5, alpha = 0.8, na.rm = T) + 
  theme_cowplot(font_size = 8, line_size = 0.25) +
  theme(legend.position = "bottom") +
  scale_color_manual(values = dissociationcolor) +
  scale_x_continuous(name="log fold difference",
                      limits = c(-10, 10)) +
  scale_y_continuous(name="-log10 p-value") +
  geom_hline(yintercept = 1,  size = 0.25, linetype = 2 ) + 
  scale_shape_manual(values = c(1,16,16))  +
  labs(color = "Overexpressed in", shape = "Overexpressed in")
volcanoplot  

  
pdf(file = "../figures/02_volcanoplots/Treatment_volcano_scalesfree.pdf", width=3, height=3)
plot(volcanoplot)
dev.off()
```

Caption: Differntial gene expression according to treatment is asymetric, with more genes enrighted in DISS. only 3 canddiate learning and memory genes identified.


Plotting CA1 vs. DG volcano plots. The color here is set inside. 

```{r subfield}
volcanoplot2 <- ggplot(subfield, aes(x = lfc, y = pvalue)) + 
  geom_point(aes(color = factor(direction)), 
             size = 1.5, alpha = 0.8, na.rm = T) + 
  theme_cowplot(font_size = 8, line_size = 0.25) +
  theme(legend.title=element_blank(),
        legend.position= "bottom",
        panel.grid.minor=element_blank(),
        panel.grid.major=element_blank()) + 
  scale_color_manual(values = c("DG" = "#d95f02",
                                "CA1" = "#7570b3",
                                "neither" = "#d9d9d9"),
                     breaks = c("DG", "CA1","neither"),
                     labels = c("DG (222)", 
                                "CA1 (262)", 
                                "neither"))+
  scale_x_continuous(name="log fold difference",
                     limits = c(-10, 10)) +
  scale_y_continuous(name="-log10 p-value",
                     limits = c(0, 18),
                     breaks = c(1,6,12,18)) +
  geom_hline(yintercept = 1,  size = 0.25, linetype = 2 )  
volcanoplot2 

pdf(file = "../figures/02_volcanoplots/CA1DG_volcano.pdf", width=3, height=3)
plot(volcanoplot2)
dev.off()

```


## Useful R tutorials
- [ggplot axis help](http://ggplot2.tidyverse.org/reference/scale_continuous.html)
- [grepply help](http://www.gettinggeneticsdone.com/2016/01/repel-overlapping-text-labels-in-ggplot2.html)
