==> scripts/03_heatmaps.Rmd <==
---
title: "heatmaps"
output: md_document
---

```{r setup, include=FALSE}
library(pheatmap)
library(viridis)
library(dplyr)
library(stringr)

# user defined functions and options
source("heatmapfunction.R") 

# set output file for figures 
knitr::opts_chunk$set(fig.path = '../figures/03_heatmaps/')
```


```{r readfiles}
dissocation_DEGs <- read.csv("../results/heatmap_DEGs.csv", header = T, check.names = F)
dissocation_df <-read.csv("../results/heatmap_colData.csv", header = T, row.names = 1)
dissocation_df <- dissocation_df %>% dplyr::select(Subfield, Treatment)
dissocation_DEGs$rownames <- str_to_upper(dissocation_DEGs$rownames)
names(dissocation_DEGs)
head(dissocation_df)
```

Here is the function I wrote to make two heatmaps, one a png and one a pdf. The goal is to have the ~ top 30 genes that are differentially expressed according to treatment.


```{r heatmaps, fig.width=3.5, fig.height=3.25}
#colorpalette
dissocation_colors <- list(Treatment = c(HOMO = (values=c("white")), 
                                         DISS = (values=c("black"))),
                           Subfield = c(CA1 = (values=c("#7570b3")),
                                        CA3 = (values=c("#1b9e77")), 
                                        DG = (values=c("#d95f02"))))

# heatmap_png and heatmap_pdf usage: DEGs (aka expression matrix), ann_colors, df (aka annotation df), title, clustercolsmethod)

# figure 2b
heatmap_png(dissocation_DEGs, dissocation_colors, dissocation_df, " ", "euclidean")
heatmap_pdf(dissocation_DEGs, dissocation_colors, dissocation_df, " ", "euclidean")

#correlation heatmap (not used in manuscript)
heatmap_png(dissocation_DEGs, dissocation_colors, dissocation_df, " ", "correlation")
```
