==> scripts/04_heatmaps.Rmd <==
---
title: "heatmaps"
author: "Rayna M Harris"
date: "12/16/2017"
output: md_document
---

```{r setup, include=FALSE}
library(pheatmap)
library(viridis)
library(dplyr)

# set output file for figures 
knitr::opts_chunk$set(fig.path = '../figures/04_heatmaps/')
```


```{r readfiles}
dissocation_DEGs <- read.csv("../results/heatmap_DEGs.csv", header = T, check.names = F)
dissocation_df <-read.csv("../results/heatmap_colData.csv", header = T, row.names = 1)
dissocation_df <- dissocation_df %>% dplyr::select(Subfield, Treatment)
names(dissocation_DEGs)
head(dissocation_df)
```

Here is the function I wrote to make two heatmaps, one a png and one a pdf. The goal is to have the ~ top 30 genes that are differentially expressed according to treatment.


```{r heatmaps}
#colorpalette
dissocation_colors <- list(Treatment = c(HOMO = (values=c("dodgerblue2")), 
                                         DISS = (values=c("firebrick1"))),
                           Subfield = c(CA1 = (values=c("#7570b3")),
                                        CA3 = (values=c("#1b9e77")), 
                                        DG = (values=c("#d95f02"))))

# the function "Heatmaps" created a png and pdf of my desired heatmap

# usage: name of matrix, color palette, lengend, title
Heatmaps <- function(DEGs, ann_colors, df, main){
  
    myfile <-  paste("../figures/04_heatmaps/", substitute(DEGs), ".pdf", sep="")
  
  DEGs <- DEGs[order(DEGs$padjmin),]
  DEGs <- head(DEGs, 30)
  print(head(DEGs, 30))

 rownames(DEGs) <- DEGs$rownames
drop.cols <-colnames(DEGs[,grep("padj|pval|rownames", colnames(DEGs))])
DEGs <- DEGs %>% select(-one_of(drop.cols))
DEGs <- as.matrix(DEGs)
DEGs <- DEGs - rowMeans(DEGs)

  paletteLength <- 30
  myBreaks <- c(seq(min(DEGs), 0, length.out=ceiling(paletteLength/2) + 1), 
              seq(max(DEGs)/paletteLength, max(DEGs), length.out=floor(paletteLength/2)))
  
pheatmap(DEGs, show_colnames=F, show_rownames = T,
         annotation_col=df, annotation_colors = ann_colors, 
         annotation_row = NA, annotation_legend = TRUE,
         annotation_names_row = FALSE, annotation_names_col = FALSE,
         border_color = "grey60" ,
         color = viridis(30),
         clustering_method="average",
         breaks=myBreaks,
         clustering_distance_cols="correlation", 
         cluster_cols = F,
         main = main)  

pheatmap(DEGs, show_colnames=F, show_rownames = T,
         annotation_col=df, annotation_colors = ann_colors, 
         annotation_row = NA, annotation_legend = TRUE,
         annotation_names_row = FALSE, annotation_names_col = TRUE,
         treeheight_row = 0, treeheight_col = 10,
         fontsize = 7, 
         border_color = "grey60" ,
         color = viridis(30),
         width=3.5, height=3.25,
         clustering_method="average",
         breaks=myBreaks,
         clustering_distance_cols="correlation", 
         cluster_cols = F,
         main = main,
         filename =  myfile)
}

Heatmaps(dissocation_DEGs, dissocation_colors, dissocation_df, " ")
```
